+++ Source/Tools/Mangos.WoWFakeClient/RC4.cs
@@ -75,7 +75,7 @@ namespace Mangos.WoWFakeClient
                 temp = key[key[257] & 0xFF];
                 key[key[257] & 0xFF] = key[key[256] & 0xFF];
                 key[key[256] & 0xFF] = temp;
-                data[i] = data[i] ^ key[Conversions.ToInteger(key[key[257]]) + Conversions.ToInteger(key[key[256]]) & 0xFF];
+                data[i] = (byte)(data[i] ^ key[Conversions.ToInteger(key[key[257]]) + Conversions.ToInteger(key[key[256]]) & 0xFF]);
             }
         }
     }
